Rework message parsing
Message parsing is now handled by message-type specific classes, and
is a bit more foolproof than before. There is still a possible
race-condition between corroutines, which is partially solved with
locks. But if messages are received too quickly, or a keepalive or
refresh happens at the wrong time, messages can become garbled.